refactor : UserViewController 와 ViewController 통합 및 코드 순서 변경
